Handle focus on search field correctly
Fixes search field which was not handling key events corretly and
allowed key presses to bubble up to the Jupyter handler. This commit
allows to focus in the search field and handle key presses in a standard
way, to focus out, press the `<Escape>` key.

Signed-off-by: Marek Cermak <macermak@redhat.com>

modified:   jupyter_datatables/js/jupyter-datatables.js